+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Optional support for comments in .projectile dirconfig files using `projectile-dirconfig-comment-prefix`.
 * [#1497](https://github.com/bbatsov/projectile/pull/1497): New command `projectile-run-gdb` (<kbd>x g</kbd> in `projectile-command-map`).
+* Add [Bazel](https://bazel.build) project type.
 
 ### Bugs fixed